+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/docsig/compare/v0.31.0...HEAD)
 ------------------------------------------------------------------------
+### Added
+- Add `-P/--check-property-returns` argument
 
 [0.31.0](https://github.com/jshwi/docsig/releases/tag/v0.31.0) - 2022-12-11
 ------------------------------------------------------------------------